Create Aggregate_func_The_Metropolitan_Museum_of_Art
The Metropolitan Museum of Art of New York is one of the world’s largest and finest art museums.
In this project, i am working with a table named met that contains the museum’s collection of American Decorative Arts.

Some tasks I've done:

How many pieces are in the American Decorative Art collection?
Count the number of pieces where the category includes ‘celery’
Find the title and medium of the oldest piece(s) in the collection.
Find the top 10 countries with the most pieces in the collection.
There are all kinds of American decorative art in the Met’s collection. Find the categories having more than 100 pieces.
Count the number of pieces where the medium contains ‘gold’ or ‘silver’ and sort them + use 'CASE' construction.