Create Joining_Multiple_Tables_Welp
I am going to use data from an exciting app called Welp. Users love Welp because it allows them to leave reviews of businesses in their city and see how other people reviewed the businesses.For this project, I will be working with two tables: places; reviews;

Some tasks I've done:

If each dollar sign ($) represents $10, how could you find all places that cost $20 or less?
Write a query to do an INNER JOIN on the two tables to show all reviews for restaurants that have at least one review
Write a query to do a LEFT JOIN on the tables, selecting the same columns as the previous question
Write a query to find all the ids of places that currently do not have any reviews in our reviews table
Write a query using the WITH clause to select all the reviews that happened in 2020. JOIN the places to your WITH query to see a log of all reviews from 2020
Write a query that finds the reviewer with the most reviews that are BELOW the average rating for places